+++ include/boost/beast/core/impl/buffers_adaptor.hpp
@@ -113,7 +113,7 @@ class buffers_adaptor<MutableBufferSequence>::
       readable_bytes<isMutable>::
       const_iterator
 {
-    iter_type it_;
+    iter_type it_{};
     buffers_adaptor const* b_ = nullptr;
 
 public:
@@ -259,7 +259,7 @@ template<class MutableBufferSequence>
 class buffers_adaptor<MutableBufferSequence>::
 mutable_buffers_type::const_iterator
 {
-    iter_type it_;
+    iter_type it_{};
     buffers_adaptor const* b_ = nullptr;
 
 public:
@@ -442,6 +442,8 @@ buffers_adaptor<MutableBufferSequence>::
 operator=(buffers_adaptor const& other) ->
     buffers_adaptor&
 {
+    if(this == &other)
+        return *this;
     auto const nbegin = std::distance<iter_type>(
         net::buffer_sequence_begin(other.bs_),
             other.begin_);