+++ include/boost/beast/core/impl/flat_static_buffer.hpp
@@ -102,6 +102,8 @@ operator=(flat_static_buffer const& other) ->
     flat_static_buffer<N>&
 {
     using net::buffer_copy;
+    if(this == &other)
+        return *this;
     this->consume(this->size());
     this->commit(buffer_copy(
         this->prepare(other.size()), other.data()));